llmspell_testing::mocks

Struct MockTool

pub struct MockTool { /* private fields */ }

Implementations

impl MockTool

pub fn checkpoint(&mut self)

Validate that all current expectations for all methods have been satisfied, and discard them.

pub fn new() -> Self

Create a new mock object with no expectations.

This method will not be generated if the real struct already has a new method. However, it will be generated if the struct implements a trait with a new method. The trait’s new method can still be called like <MockX as TraitY>::new

impl MockTool

pub fn expect_metadata(&mut self) -> &mut Expectation

Create an Expectation for mocking the metadata method

pub fn expect_execute_impl(&mut self) -> &mut Expectation

Create an Expectation for mocking the execute_impl method

pub fn expect_validate_input(&mut self) -> &mut Expectation

Create an Expectation for mocking the validate_input method

pub fn expect_handle_error(&mut self) -> &mut Expectation

Create an Expectation for mocking the handle_error method

pub fn expect_stream_execute(&mut self) -> &mut Expectation

Create an Expectation for mocking the stream_execute method

pub fn expect_supports_streaming(&mut self) -> &mut Expectation

Create an Expectation for mocking the supports_streaming method

pub fn expect_supports_multimodal(&mut self) -> &mut Expectation

Create an Expectation for mocking the supports_multimodal method

pub fn expect_supported_media_types(&mut self) -> &mut Expectation

Create an Expectation for mocking the supported_media_types method

impl MockTool

pub fn expect_schema(&mut self) -> &mut Expectation

Create an Expectation for mocking the schema method

pub fn expect_category(&mut self) -> &mut Expectation

Create an Expectation for mocking the category method

pub fn expect_security_level(&mut self) -> &mut Expectation

Create an Expectation for mocking the security_level method

pub fn expect_validate_parameters(&mut self) -> &mut Expectation

Create an Expectation for mocking the validate_parameters method